Completed hw4
