+++ app/src/main/AndroidManifest.xml
@@ -20,14 +20,8 @@
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity
-            android:name=".AddTuringPersonActivity"
-            android:exported="true">
-
         </activity>
     </application>
